refactor(services): rename service files to shorter names
